refactor: Rename UserTableEmptyState import in UserTable component
